Fix warning

+++ python/lib/davis/dataset/utils.py
@@ -41,7 +41,8 @@ def db_statistics(per_frame_values):
 
 	N_bins = 4
 	#ids = np.round(np.linspace(1, len(per_frame_values),N_bins+1)).astype(np.uint)-1
-	ids = np.round(np.linspace(1,len(per_frame_values)-1,N_bins+1)).astype(np.uint)-1;
+	ids = np.round(np.linspace(1,len(per_frame_values)-1,N_bins+1))-1;
+	ids = ids.astype(np.uint8)
 
 	D_bins = [per_frame_values[ids[i]:ids[i+1]+1] for i in range(0,4)]
 	D      = np.nanmean(D_bins[0])-np.nanmean(D_bins[3])
